main-jby-fix: Ajuste id cliente de cuentas

+++ banking_backend/account-service/src/main/java/com/tcs/account_service/domain/Entity/Cuenta.java
@@ -37,7 +37,7 @@ public class Cuenta {
     private Estado estado;
 
     @Column(name = "cliente_id", nullable = false)
-    private int clienteId;
+    private Long clienteId;
 
     @PrePersist
     public void prePersist() {
