+++ banking_backend/account-service/src/main/java/com/tcs/account_service/dto/response/CuentaResponseDTO.java
@@ -14,6 +14,6 @@ public class CuentaResponseDTO {
     private TipoCuenta tipoCuenta;
     private BigDecimal saldoInicial;
     private Estado estado;
-    private int clienteId;
+    private Long clienteId;
 
 }
